feat/2 user

+++ src/main/java/com/homeis/user/controller/UserController.java
@@ -3,6 +3,7 @@ package com.homeis.user.controller;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -30,6 +31,25 @@ public class UserController {
 	private final UserService userService;
 	private final JWTUtil jwtUtil;
 	
+	@GetMapping("/{userId}")
+	public ResponseEntity<?> idExist(@PathVariable("userId") String userId) {
+		int isDup = userService.idExist(userId);
+		// id가 중복일 때
+		if (isDup == 1)	return ResponseEntity.status(HttpStatus.CONFLICT).body("id 중복!");
+		
+		return ResponseEntity.ok("생성가능!");
+	}
+	
+	@PostMapping("/register")
+	public ResponseEntity<?> register(@RequestBody User userInfo) {
+		
+		int result = userService.register(userInfo);
+		if (result != 1) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body("잘못된 접근");
+		}
+		return ResponseEntity.ok(result);
+	}
+	
 	//로그인
 	@PostMapping("/login")
 	public ResponseEntity<?> login(@RequestBody User loginInfo){
@@ -41,24 +61,12 @@ public class UserController {
 		return ResponseEntity.ok(token);
 	}
 	
-	@PostMapping("/register")
-	public ResponseEntity<Integer> register(@RequestBody User userInfo) {
-		
-		int flag = userService.register(userInfo);
-		if (flag != 1) {
-			return ResponseEntity.status(404).body(0);
-		}
-		return ResponseEntity.status(200).body(1);
-	}
-	
-	@GetMapping("/{userId}")
-	public ResponseEntity<Integer> idExist(@PathVariable("userId") String userId) {
-		User user = userService.idExist(userId);
-		if (user == null) {
-			return ResponseEntity.status(404).body(0);
-		}
-		return ResponseEntity.status(200).body(1);
-	}
+	// 관심지역 등록 보류
+//	@PostMapping("/interest-area")
+//	public ResponseEntity<?> registArea (@RequestHeader("Authorization") String tokenHeader,
+//			@RequestBody DongCode dongCode) {
+//		jwtUtil.getIdFromToken(tokenHeader.substring(7));
+//	}
 	
 	@GetMapping("/mypage/{id}")
 	public ResponseEntity<?> getUserInfo(@PathVariable("id") String id, @RequestHeader("Authorization") String tokenHeader) {
@@ -74,30 +82,54 @@ public class UserController {
 		return ResponseEntity.ok(user);
 	}
 	
-	@PutMapping
-	public ResponseEntity<Integer> updateUserInfo(@ModelAttribute User userInfo) {
-		int flag = userService.updateUserInfo(userInfo);
-		if (flag == 0) {
-			return ResponseEntity.status(404).body(0);
+	@PutMapping("/")
+	public ResponseEntity<?> updateUserInfo(@RequestBody User userInfo, @RequestHeader("Authorization") String tokenHeader) {
+		//토큰이 유효하지 않은 경우
+		boolean isValid = jwtUtil.isValid(tokenHeader.substring(7));
+		if(!isValid) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰 만료, 잘못된 접근.");
+		
+		String userId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		if(!userId.equals(userInfo.getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("잘못된 접근입니다.");
+		
+		int result = userService.updateUserInfo(userInfo);
+		if (result == 0) {
+			return ResponseEntity.status(404).build();
 		}
-		return ResponseEntity.status(200).body(1);
+		return ResponseEntity.ok(result);
 	}
 	
-	@DeleteMapping("/{userId}")
-	public ResponseEntity<Integer> deleteUser(@PathVariable("userId") String userId) {
-		int flag = userService.deleteUser(userId);
-		if (flag == 0) {
-			return ResponseEntity.status(404).body(0);
+	@DeleteMapping("/{id}")
+	public ResponseEntity<?> deleteUser(@PathVariable("id") String id, @RequestHeader("Authorization") String tokenHeader) {
+		
+		//토큰이 유효하지 않은 경우
+		boolean isValid = jwtUtil.isValid(tokenHeader.substring(7));
+		if(!isValid) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰 만료, 잘못된 접근.");
+		
+		String userId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		if(!userId.equals(id)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("잘못된 접근입니다.");
+		
+		int result = userService.deleteUser(userId);
+		if (result == 0) {
+			return ResponseEntity.status(404).build();
 		}
-		return ResponseEntity.status(200).body(1);
+		return ResponseEntity.ok(result);
 	}
 	
-	@GetMapping("/interest-area/{userId}")
-	public ResponseEntity<List<DongCode>> getInterestArea(@PathVariable("userId") String userId) {
+	@GetMapping("/interest-area/{id}")
+	public ResponseEntity<?> getInterestArea(@PathVariable("id") String id,  @RequestHeader("Authorization") String tokenHeader) {
+		//토큰이 유효하지 않은 경우
+		boolean isValid = jwtUtil.isValid(tokenHeader.substring(7));
+		if(!isValid) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰 만료, 잘못된 접근.");
+		
+		String userId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		if(!userId.equals(id)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("잘못된 접근입니다.");
+		
+		
 		List<DongCode> interestAreaList = userService.getInterestArea(userId);
 		if (interestAreaList.isEmpty()) {
 			return ResponseEntity.status(404).body(null);
 		}
 		return ResponseEntity.status(200).body(interestAreaList);
 	}
+	
 }
